services/pictures/requirements.txt
Pillow
pyheif
opencv-python
fastapi
uvicorn
redis
httpx
PyJWT[crypto]
cryptography